# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreatorProfile {
  id: string;
  displayName: string;
  age: number;
  primaryCategory: string;
  accountType: string;
  followerCount: number;
  followerTier: string;
  contentFormats: string[];
  postFrequencyPerWeek: number;
  targetAudience: string;
  aspiration: string;
  biggestChallenge: string;
  baselineEngagementRate: number;
  baselineReach: number;
  contentTone: string;
  inspiredBy: string[];
  pastTacticsTried: string[];
  postingTimeHabit: string;
  creatingSince: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreatorProfileResponse {
  profile: CreatorProfile | null;
}

export interface UpsertCreatorProfileRequest {
  displayName: string;
  /** @minimum 13 */
  age: number;
  primaryCategory: string;
  accountType: string;
  /** @minimum 0 */
  followerCount: number;
  contentFormats: string[];
  /** @minimum 0 */
  postFrequencyPerWeek: number;
  targetAudience: string;
  aspiration: string;
  biggestChallenge: string;
  baselineEngagementRate: number;
  /** @minimum 0 */
  baselineReach: number;
  contentTone: string;
  inspiredBy: string[];
  pastTacticsTried: string[];
  postingTimeHabit: string;
  creatingSince: string;
}

export interface MetricsInput {
  reach: number;
  impressions: number;
  engagementRate: number;
  followerChange: number;
  saves: number;
  shares: number;
  profileVisits: number;
  linkClicks: number;
  contentFormat: string;
  postTopic: string;
}

export interface CreateAnalysisRequest {
  metrics: MetricsInput;
  context?: string;
}

export interface CreateScreenshotInsightsRequest {
  imageDataUrl: string;
  notes?: string;
}

export type SuggestionStatus =
  (typeof SuggestionStatus)[keyof typeof SuggestionStatus];

export const SuggestionStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface Suggestion {
  id: string;
  analysisId: string;
  title: string;
  rationale: string;
  action: string;
  status: SuggestionStatus;
}

export interface Analysis {
  id: string;
  profileId: string;
  source: string;
  metrics: MetricsInput;
  summary: string;
  brutallyHonestTake: string;
  whyItHappened: string[];
  clarifyingQuestions: string[];
  nextContentPlan: string[];
  suggestions: Suggestion[];
  createdAt: string;
}

export type SubmitFeedbackRequestStatus =
  (typeof SubmitFeedbackRequestStatus)[keyof typeof SubmitFeedbackRequestStatus];

export const SubmitFeedbackRequestStatus = {
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface SubmitFeedbackRequest {
  suggestionId: string;
  status: SubmitFeedbackRequestStatus;
  reason?: string;
}

export interface CreatorMatch {
  id: string;
  handle: string;
  category: string;
  followerTier: string;
  whyMatch: string;
  growthSignal: string;
  tacticToStudy: string;
}

export interface DashboardResponse {
  profileComplete: boolean;
  analysesCount: number;
  acceptedSuggestions: number;
  rejectedSuggestions: number;
  averageEngagementRate: number;
  latestAnalysis: Analysis | null;
  calibrationNotes: string[];
}