# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MetricsInput } from "./metricsInput";
import type { Suggestion } from "./suggestion";

export interface Analysis {
  id: string;
  profileId: string;
  source: string;
  metrics: MetricsInput;
  summary: string;
  brutallyHonestTake: string;
  whyItHappened: string[];
  clarifyingQuestions: string[];
  nextContentPlan: string[];
  suggestions: Suggestion[];
  createdAt: string;
}