# lib/api-zod/src/generated/types/createAnalysisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MetricsInput } from "./metricsInput";

export interface CreateAnalysisRequest {
  metrics: MetricsInput;
  context?: string;
}