# lib/api-zod/src/generated/types/createScreenshotInsightsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateScreenshotInsightsRequest {
  imageDataUrl: string;
  notes?: string;
}