# lib/api-zod/src/generated/types/creatorMatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatorMatch {
  id: string;
  handle: string;
  category: string;
  followerTier: string;
  whyMatch: string;
  growthSignal: string;
  tacticToStudy: string;
}