# lib/api-zod/src/generated/types/creatorProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatorProfile {
  id: string;
  displayName: string;
  age: number;
  primaryCategory: string;
  accountType: string;
  followerCount: number;
  followerTier: string;
  contentFormats: string[];
  postFrequencyPerWeek: number;
  targetAudience: string;
  aspiration: string;
  biggestChallenge: string;
  baselineEngagementRate: number;
  baselineReach: number;
  contentTone: string;
  inspiredBy: string[];
  pastTacticsTried: string[];
  postingTimeHabit: string;
  creatingSince: string;
  createdAt: string;
  updatedAt: string;
}