# lib/api-zod/src/generated/types/creatorProfileResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreatorProfile } from "./creatorProfile";

export interface CreatorProfileResponse {
  profile: CreatorProfile | null;
}