# lib/api-zod/src/generated/types/dashboardResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Analysis } from "./analysis";

export interface DashboardResponse {
  profileComplete: boolean;
  analysesCount: number;
  acceptedSuggestions: number;
  rejectedSuggestions: number;
  averageEngagementRate: number;
  latestAnalysis: Analysis | null;
  calibrationNotes: string[];
}