# lib/api-zod/src/generated/types/metricsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MetricsInput {
  reach: number;
  impressions: number;
  engagementRate: number;
  followerChange: number;
  saves: number;
  shares: number;
  profileVisits: number;
  linkClicks: number;
  contentFormat: string;
  postTopic: string;
}