# lib/api-zod/src/generated/types/submitFeedbackRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubmitFeedbackRequestStatus } from "./submitFeedbackRequestStatus";

export interface SubmitFeedbackRequest {
  suggestionId: string;
  status: SubmitFeedbackRequestStatus;
  reason?: string;
}