# lib/api-zod/src/generated/types/suggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SuggestionStatus } from "./suggestionStatus";

export interface Suggestion {
  id: string;
  analysisId: string;
  title: string;
  rationale: string;
  action: string;
  status: SuggestionStatus;
}