# lib/api-zod/src/generated/types/upsertCreatorProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpsertCreatorProfileRequest {
  displayName: string;
  /** @minimum 13 */
  age: number;
  primaryCategory: string;
  accountType: string;
  /** @minimum 0 */
  followerCount: number;
  contentFormats: string[];
  /** @minimum 0 */
  postFrequencyPerWeek: number;
  targetAudience: string;
  aspiration: string;
  biggestChallenge: string;
  baselineEngagementRate: number;
  /** @minimum 0 */
  baselineReach: number;
  contentTone: string;
  inspiredBy: string[];
  pastTacticsTried: string[];
  postingTimeHabit: string;
  creatingSince: string;
}